surya

+++ commands/owoify.js
@@ -7,7 +7,10 @@ module.exports = {
         if (message.author.bot) return;
         const args2 = message.content.trim()
         let owo = owoify(args2)
-        message.delete()            
-        return message.channel.send(owo).catch(e => console.log(e)) 
+
+        if(message.author.username === 'sushirymushia'){
+            message.delete()            
+            return message.channel.send(owo).catch(e => console.log(e)) 
+        }
     }
 }
